Support link-generation functions again
`cl-pushnew' is pushes an element onto the list if it is not already
present.  Unfortunately, this does not account for alists; since there
was no member (link . apiwrap-stdgenlink), it was always pushed to the
front of the list for `alist-get' to find.

Using standard functions makes it a little clearer what's going on and
fixes this issue.