Fix issue with override functions not being respected
Component functions are provided has plists but processed as
alists (converted with `apiwrap-plist->alist').  In -genmacro, we
mixed up these two types by `append'ing them together.  This caused
`alist-get' to consistently miss those functions.

The resolution is to provide another parameter to -gendefun for
'override' functions that can be dealt with as their own entity (for
syntax resolution) and converted into an alist before prepending onto
the standard functions.