Make parameter resolution less hacky
Rename

    `apiwrap-resolve-api-params'

to

    `apiwrap-genform-resolve-api-params'

Functionally relevant: instead of building a sexp with text insertion
coupled with `read', we're building the list as data. This should be
much less error-prone and, well, downright ugly.

There is yet a little hackiness involved in macro expansion and
argument evaluation, but that's a harder nut to crack.