Add optional parameters for pre-process functions
`:pre-process-{params,data}' are now available to customize how a
wrapped resource handles its natural input.

For example, if a resource expects a list of IDs, but the natural
input would be a list of objects, you can use a function as one of
these parameters to extract the IDs from the objects.  This list of
IDs is then passed to the `:request' function instead of the original
list of objects your wrapper was called with.